# docs-site/docs/api/openapi/coinline.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Host Platform API",
    "version": "v1"
  },
  "paths": {
...
    "/api/downloads/{id}": {
      "get": {
        "tags": [
          "Downloads"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/downloads/{id}/cancel": {
      "post": {
        "tags": [
          "Downloads"
        ],
        "summary": "Cancels an in-flight orchestration batch. Requires `?confirm=true`. Writes an audit event.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/downloads/{id}/retry": {
      "post": {
        "tags": [
          "Downloads"
        ],
        "summary": "Queues a replacement batch after failure/cancel. Requires `?confirm=true`. Writes an audit event.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}